+++ src/array.ml
@@ -223,6 +223,7 @@ module type Permissioned = sig
   val of_list_rev_map : 'a list -> f:('a -> 'b) -> ('b, [< _ perms]) t
   val of_list_rev_mapi : 'a list -> f:(int -> 'a -> 'b) -> ('b, [< _ perms]) t
   val replace : ('a, [> read_write]) t -> int -> f:('a -> 'a) -> unit
+  [@@deprecated "[since 2018-09] use [t.(i) <- f (t.(i))] instead"]
   val replace_all : ('a, [> read_write]) t -> f:('a -> 'a) -> unit
   [@@deprecated "[since 2018-03] use [map_inplace] instead"]
   val map_inplace : ('a, [> read_write]) t -> f:('a -> 'a) -> unit
@@ -374,6 +375,7 @@ module type S = sig
   val of_list_rev_map : 'a list -> f:('a -> 'b) -> 'b t
   val of_list_rev_mapi : 'a list -> f:(int -> 'a -> 'b) -> 'b t
   val replace : 'a t -> int -> f:('a -> 'a) -> unit
+  [@@deprecated "[since 2018-09] use [t.(i) <- f (t.(i))] instead"]
   val replace_all : 'a t -> f:('a -> 'a) -> unit
   [@@deprecated "[since 2018-03] use [map_inplace] instead"]
   val map_inplace : 'a t -> f:('a -> 'a) -> unit