+++ src/array.mli
@@ -235,6 +235,7 @@ module Permissioned : sig
   val of_list_rev_map : 'a list -> f:('a -> 'b) -> ('b, [< _ perms]) t
   val of_list_rev_mapi : 'a list -> f:(int -> 'a -> 'b) -> ('b, [< _ perms]) t
   val replace     : ('a, [> read_write]) t -> int -> f:('a -> 'a) -> unit
+  [@@deprecated "[since 2018-09] use [t.(i) <- f (t.(i))] instead"]
   val replace_all : ('a, [> read_write]) t -> f:('a -> 'a) -> unit
   [@@deprecated "[since 2018-03] use [map_inplace] instead"]
   val map_inplace  : ('a, [> read_write]) t -> f:('a -> 'a) -> unit