+++ CHANGES.md
@@ -20,6 +20,8 @@ you might want to have a look at Base's
 - Deprecate `Timing_wheel_ns.alarm_upper_bound` in favor of
   `Timing_wheel_ns.max_allowed_alarm_time`.
 
+- Deprecated `Array.replace arr i ~f` in favor of using `arr.(i) <- (f (arr.(i)))`
+
 ## v0.11
 
 - Add `Bigstring.unsafe_resize` to allow reallocating in place.